backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
sqlite3  # Built into Python standard library

# OpenAI integration
openai==1.3.7

# Image processing
Pillow==10.1.0

# Environment variables
python-dotenv==1.0.0

# CORS support
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# Data validation
pydantic==2.5.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
